+++ src/test/java/com/salesforce/storm/spout/dynamic/DynamicSpoutTest.java
@@ -57,8 +57,8 @@ import com.tngtech.java.junit.dataprovider.UseDataProvider;
 import org.apache.storm.generated.StreamInfo;
 import com.google.common.base.Charsets;
 import org.apache.storm.task.TopologyContext;
-import org.apache.storm.topology.IRichSpout;
 import org.apache.storm.topology.OutputFieldsGetter;
+import org.apache.storm.tuple.Fields;
 import org.apache.storm.utils.Utils;
 import org.apache.zookeeper.KeeperException;
 import org.junit.Before;
@@ -72,7 +72,6 @@ import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.time.Clock;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Iterator;
@@ -1051,12 +1050,13 @@ public class DynamicSpoutTest {
      * declareOutputFields() method with default to using 'default' stream.
      */
     @Test
-    public void testDeclareOutputFields_without_stream() {
+    @UseDataProvider("provideOutputFields")
+    public void testDeclareOutputFields_without_stream(final Object inputFields, final String[] expectedFields) {
         // Create config with null stream id config option.
         final Map<String,Object> config = getDefaultConfig("SidelineSpout-", null);
 
         // Define our output fields as key and value.
-        config.put(SpoutConfig.OUTPUT_FIELDS, "key,value");
+        config.put(SpoutConfig.OUTPUT_FIELDS, inputFields);
 
         final OutputFieldsGetter declarer = new OutputFieldsGetter();
 
@@ -1072,7 +1072,7 @@ public class DynamicSpoutTest {
         assertTrue(fieldsDeclaration.containsKey(Utils.DEFAULT_STREAM_ID));
         assertEquals(
             fieldsDeclaration.get(Utils.DEFAULT_STREAM_ID).get_output_fields(),
-            Lists.newArrayList("key", "value")
+            Lists.newArrayList(expectedFields)
         );
 
         spout.close();
@@ -1083,12 +1083,13 @@ public class DynamicSpoutTest {
      * in the declareOutputFields() method.
      */
     @Test
-    public void testDeclareOutputFields_with_stream() {
+    @UseDataProvider("provideOutputFields")
+    public void testDeclareOutputFields_with_stream(final Object inputFields, final String[] expectedFields) {
         final String streamId = "foobar";
         final Map<String,Object> config = getDefaultConfig("SidelineSpout-", streamId);
 
         // Define our output fields as key and value.
-        config.put(SpoutConfig.OUTPUT_FIELDS, "key,value");
+        config.put(SpoutConfig.OUTPUT_FIELDS, inputFields);
 
         final OutputFieldsGetter declarer = new OutputFieldsGetter();
 
@@ -1103,10 +1104,33 @@ public class DynamicSpoutTest {
         assertTrue(fieldsDeclaration.containsKey(streamId));
         assertEquals(
             fieldsDeclaration.get(streamId).get_output_fields(),
-            Lists.newArrayList("key", "value")
+            Lists.newArrayList(expectedFields)
         );
 
         spout.close();
+    }
+
+    /**
+     * Provides various inputs to be split.
+     */
+    @DataProvider
+    public static Object[][] provideOutputFields() throws InstantiationException, IllegalAccessException {
+        return new Object[][] {
+            // String inputs, these get split and trimmed.
+            { "key,value", new String[] {"key", "value"} },
+            { "key, value", new String[] {"key", "value"} },
+            { " key    , value  ,", new String[] {"key", "value"} },
+
+            // List of Strings, used as is.
+            { Lists.newArrayList("key", "value"), new String[] { "key", "value"} },
+            { Lists.newArrayList("  key  ", " value"), new String[] { "  key  ", " value"} },
+            { Lists.newArrayList("key,value", "another"), new String[] { "key,value", "another"} },
+
+            // Fields inputs, used as is.
+            { new Fields("key", "value"), new String[] { "key", "value" } },
+            { new Fields(" key ", "    value"), new String[] { " key ", "    value" } },
+            { new Fields("key,value ", "another"), new String[] { "key,value ", "another" } },
+        };
     }
 
     /**